Fix big-endian support
Main issue was endian conversion of encrypted data. When data is
stored as encrypted it is stored as array of uint32_t LE. When
this is loaded on BE system it needs to be converted to BE before
decrypt attempt. After decrypt if data is further casted to type
which consist of values with different size than 32 bits (i.e.
mpq_hash_s which has uint16_t fields) it needs to be converted back
to the format endian (LE), casted and after that, converted to
platform endian (BE).